chore(aws): Add `AWSDebuggable` mixin
Moves debug-friendly `toString` logic to a new type to allow classes to opt-in to the `toString` behavior separately from the serialization/equatable behaviors we have. For example, AWS credentials should be serializable and equatable, but by default `toString` should conceal sensitive info. This was done before my manually overriding `toString`, but this is error prone and the new system makes it clear which types are "debuggable".

commit-id:faf596b4